requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tenacity==9.1.4
tzdata==2025.3
urllib3==2.6.3
